Fix type issues in rustmath-groups
1. Add Hash derive to IndexedFreeAbelianGroupElement
   - Added Hash to derive macro on line 482 of indexed_free_group.rs

2. Fix i32→isize conversions throughout groups module
   - Changed FinitelyPresentedGroup::relations from Vec<Vec<i32>> to Vec<Vec<isize>>
   - Changed FinitelyPresentedGroupElement::word from Vec<i32> to Vec<isize>>
   - Changed RewritingSystem::rules from Vec<(Vec<i32>, Vec<i32>)> to Vec<(Vec<isize>, Vec<isize>)>
   - Updated all method signatures to use isize instead of i32
   - Removed unnecessary .map(|x| x as i32) conversions in finitely_presented_named.rs
   - Fixed type conversions in artin.rs

3. Fix Complex::from() calls to use Complex::new()
   - Replaced all Complex::from(value) with Complex::new(value, 0.0) in class_function.rs
   - Complex type does not have a From<f64> implementation, needs explicit real/imaginary parts

+++ rustmath-groups/src/artin.rs
@@ -224,7 +224,7 @@ impl ArtinGroup {
             }
         }
 
-        let fp_group = FinitelyPresentedGroup::new(generator_names.clone(), relations.iter().map(|r| r.word().iter().flat_map(|(gen, exp)| vec![*gen as i32, *exp as i32]).collect()).collect());
+        let fp_group = FinitelyPresentedGroup::new(generator_names.clone(), relations.iter().map(|r| r.word().iter().flat_map(|(gen, exp)| vec![*gen, *exp]).collect()).collect());
 
         Self {
             coxeter_matrix,
